Refactor to not repeat code, add columnLayout folder
